# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AdvertisementRequest:
    properties:
      content_cta_url:
        type: string
      content_image_url:
        type: string
      display_duration:
        type: integer
      end_at:
        type: string
      location_id:
        type: string
      navigate:
        type: string
      placement:
        type: string
      reward_amount:
        type: integer
      reward_type:
        type: string
      size_height:
        type: integer
      size_width:
        type: integer
      start_at:
        type: string
      type:
        type: string
    type: object
  dto.AdvertisementResponse:
    properties:
      content_cta_url:
        type: string
      content_image_url:
        type: string
      created_at:
        type: string
      display_duration:
        type: integer
      end_at:
        type: string
      id:
        type: string
      location_id:
        type: string
      navigate:
        type: string
      placement:
        type: string
      published_at:
        type: string
      reward_amount:
        type: integer
      reward_type:
        type: string
      size_height:
        type: integer
      size_width:
        type: integer
      start_at:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
      venue_id:
        type: string
    type: object
  dto.AppCode:
    enum:
    - 0
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    type: integer
    x-enum-comments:
      CodeAuthRequired: Request requires authentication; no valid token provided.
      CodeBusinessLogic: Valid request but invalid in the current context.
      CodeConflict: Duplicate resource or incompatible state.
      CodeInternalError: Unexpected server-side error.
      CodeNotFound: Requested resource does not exist.
      CodePermissionDenied: Authenticated but lacks permission to perform this action.
      CodeRateLimited: Too many requests; client should back off.
      CodeSuccess: Request processed successfully.
      CodeValidationError: 'Invalid input: missing field, wrong type, or format violation.'
    x-enum-descriptions:
    - Request processed successfully.
    - 'Invalid input: missing field, wrong type, or format violation.'
    - Request requires authentication; no valid token provided.
    - Authenticated but lacks permission to perform this action.
    - Requested resource does not exist.
    - Duplicate resource or incompatible state.
    - Valid request but invalid in the current context.
    - Too many requests; client should back off.
    - Unexpected server-side error.
    x-enum-varnames:
    - CodeSuccess
    - CodeValidationError
    - CodeAuthRequired
    - CodePermissionDenied
    - CodeNotFound
    - CodeConflict
    - CodeBusinessLogic
    - CodeRateLimited
    - CodeInternalError
...
  dto.EventLogResponse:
    properties:
      created_at:
        type: string
      device_id:
        type: string
      id:
        type: string
      ip_address:
        type: string
      name:
        type: string
      params:
        type: object
      user_agent:
        type: string
      user_id:
        type: string
      venue_id:
        type: string
    type: object
...
  dto.PaginatedData:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
  dto.PaginationMeta:
    properties:
      current_page:
        type: integer
      has_next_page:
        type: boolean
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  dto.Response:
    properties:
      code:
        $ref: '#/definitions/dto.AppCode'
      data: {}
      messages:
        items:
          type: string
        type: array
    type: object
...
  dto.TrackEventRequest:
    properties:
      device_id:
        type: string
      name:
        type: string
      params:
        type: object
      user_id:
        type: string
    required:
    - name
    type: object
  dto.TrackSearchRequest:
    properties:
      device_id:
        type: string
      term:
        type: string
      user_id:
        type: string
    required:
    - term
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Digimap backend REST API
  title: Digimap Backend API
  version: "1.0"
paths:
...
  /public/venues/{id}/events:
    post:
      consumes:
      - application/json
      description: Track an analytics event (no auth required)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Event data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TrackEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.EventLogResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Track event
      tags:
      - analytics
  /public/venues/{id}/searches:
    post:
      consumes:
      - application/json
      description: Track a search query (no auth required)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Search data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TrackSearchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Track search
      tags:
      - analytics
...
  /venues/{id}/ads:
    get:
      description: List advertisements for a venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List advertisements
      tags:
      - ads
    post:
      consumes:
      - application/json
      description: Create a new advertisement (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Advertisement details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdvertisementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdvertisementResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create advertisement
      tags:
      - ads
  /venues/{id}/ads/{adID}:
    delete:
      description: Delete an advertisement (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Advertisement ID
        in: path
        name: adID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete advertisement
      tags:
      - ads
    get:
      description: Get an advertisement by ID
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Advertisement ID
        in: path
        name: adID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdvertisementResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get advertisement
      tags:
      - ads
    put:
      consumes:
      - application/json
      description: Update an advertisement (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Advertisement ID
        in: path
        name: adID
        required: true
        type: string
      - description: Advertisement details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdvertisementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdvertisementResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update advertisement
      tags:
      - ads
  /venues/{id}/ads/{adID}/publish:
    post:
      description: Publish a draft advertisement (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Advertisement ID
        in: path
        name: adID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Publish advertisement
      tags:
      - ads
  /venues/{id}/analytics/events:
    get:
      description: List analytics event logs for a venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List event logs
      tags:
      - analytics
  /venues/{id}/analytics/searches:
    get:
      description: List search query logs for a venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List search queries
      tags:
      - analytics
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"